Share WAR layer construction code
