+++ jib-gradle-plugin/src/test/java/com/google/cloud/tools/jib/gradle/GradleLayerConfigurationsTest.java
@@ -303,7 +303,6 @@ public class GradleLayerConfigurationsTest {
             webAppDirectory.resolve("jib-exploded-war/WEB-INF/lib/dependencyX-1.0.0-SNAPSHOT.jar"));
     ImmutableList<Path> expectedResourcesFiles =
         ImmutableList.of(
-            webAppDirectory.resolve("jib-exploded-war/META-INF"),
             webAppDirectory.resolve("jib-exploded-war/META-INF/context.xml"),
             webAppDirectory.resolve("jib-exploded-war/Test.jsp"),
             webAppDirectory.resolve("jib-exploded-war/WEB-INF/classes/empty_dir"),
@@ -312,13 +311,11 @@ public class GradleLayerConfigurationsTest {
     ImmutableList<Path> expectedClassesFiles =
         ImmutableList.of(
             webAppDirectory.resolve("jib-exploded-war/WEB-INF/classes/HelloWorld.class"),
+            webAppDirectory.resolve("jib-exploded-war/WEB-INF/classes/empty_dir"),
             webAppDirectory.resolve("jib-exploded-war/WEB-INF/classes/package/Other.class"));
     ImmutableList<Path> expectedExtraFiles =
         ImmutableList.of(
-            extraFilesDirectory.resolve("a"),
-            extraFilesDirectory.resolve("a/b"),
             extraFilesDirectory.resolve("a/b/bar"),
-            extraFilesDirectory.resolve("c"),
             extraFilesDirectory.resolve("c/cat"),
             extraFilesDirectory.resolve("foo"));
 
@@ -338,7 +335,6 @@ public class GradleLayerConfigurationsTest {
         configuration.getSnapshotDependencyLayerEntries());
     assertExtractionPathsUnordered(
         Arrays.asList(
-            "/my/app/META-INF",
             "/my/app/META-INF/context.xml",
             "/my/app/Test.jsp",
             "/my/app/WEB-INF/classes/empty_dir",
@@ -348,11 +344,11 @@ public class GradleLayerConfigurationsTest {
     assertExtractionPathsUnordered(
         Arrays.asList(
             "/my/app/WEB-INF/classes/HelloWorld.class",
+            "/my/app/WEB-INF/classes/empty_dir",
             "/my/app/WEB-INF/classes/package/Other.class"),
         configuration.getClassLayerEntries());
     assertExtractionPathsUnordered(
-        Arrays.asList("/a", "/a/b", "/a/b/bar", "/c", "/c/cat", "/foo"),
-        configuration.getExtraFilesLayerEntries());
+        Arrays.asList("/a/b/bar", "/c/cat", "/foo"), configuration.getExtraFilesLayerEntries());
   }
 
   @Test
@@ -374,7 +370,6 @@ public class GradleLayerConfigurationsTest {
         configuration.getSnapshotDependencyLayerEntries());
     assertExtractionPathsUnordered(
         Arrays.asList(
-            "/jetty/webapps/ROOT/META-INF",
             "/jetty/webapps/ROOT/META-INF/context.xml",
             "/jetty/webapps/ROOT/Test.jsp",
             "/jetty/webapps/ROOT/WEB-INF/classes/empty_dir",
@@ -384,11 +379,11 @@ public class GradleLayerConfigurationsTest {
     assertExtractionPathsUnordered(
         Arrays.asList(
             "/jetty/webapps/ROOT/WEB-INF/classes/HelloWorld.class",
+            "/jetty/webapps/ROOT/WEB-INF/classes/empty_dir",
             "/jetty/webapps/ROOT/WEB-INF/classes/package/Other.class"),
         configuration.getClassLayerEntries());
     assertExtractionPathsUnordered(
-        Arrays.asList("/a", "/a/b", "/a/b/bar", "/c", "/c/cat", "/foo"),
-        configuration.getExtraFilesLayerEntries());
+        Arrays.asList("/a/b/bar", "/c/cat", "/foo"), configuration.getExtraFilesLayerEntries());
   }
 
   @Test
